requirements.txt
bs4
knackpy
requests
flask
flask_cors
pytest
pyyaml
zappa